yandex-kafka/requirements.txt
aiohttp==3.8.0
aiohttp-rest-api==0.0.6
kafka-python==2.0.2
requests==2.22.0
uuid==1.30
